requirements.txt
# Core ML/CV
onnxruntime>=1.16.0
opencv-python>=4.8.0
numpy>=1.24.0
ultralytics>=8.0.0

# Data handling
pandas>=2.0.0
pyarrow>=14.0.0
pydantic>=2.0.0
pyyaml>=6.0.0

# Tracking
filterpy>=1.4.5
scipy>=1.11.0
lap>=0.4.0

# Geometry
shapely>=2.0.0

# UI/CLI
streamlit>=1.28.0
typer>=0.9.0
rich>=13.0.0

# Logging
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code quality
ruff>=0.1.0

